requirements.txt
PyMuPDF==1.23.8
pdfplumber==0.10.3
pdfminer.six==20221105
pytesseract==0.3.10
Pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2
scipy==1.16.1
threadpoolctl  ==    3.6.0
pypdfium2     ==     4.30.0
cffi            ==   1.17.1
charset-normalizer== 3.4.2
cryptography       ==45.0.5
packaging         == 25.0
pycparser       ==   2.22

